additional refactoring
